feat: implement PokeAPI integration service with caching and Pydantic schema validation
