+++ backend/requirements.txt
@@ -5,3 +5,4 @@ pytesseract
 Pillow
 httpx
 supabase
+async-lru
