+++ backend/schemas/pokemon.py
@@ -1,0 +1,14 @@
+from pydantic import BaseModel
+from typing import List, Optional, Dict
+
+class PokemonInfo(BaseModel):
+    id: int
+    name: str                   
+    english_name: str           
+    types: List[str]            
+    abilities: List[str]        
+    base_stats: Dict[str, int]  
+    weight_kg: float            
+    height_m: float             
+    moves: List[str]            
+    image_url: Optional[str] = None
